requirements.txt
pyserial==3.5
typing-extensions==4.8.0
